feat/refactor: 1. Change Category rework 2 Side panel new options 3. Modal refactor

+++ src/containers/User/Settings/DeleteAccount.tsx
@@ -9,6 +9,7 @@ const DeleteAccount = () => {
   const { user } = useSession();
   const router = useRouter();
   const handleSubmit = async () => {
+    if (!user?.username) return;
     const res = await fetch(`/api/users/${user?.username}/`, {
       method: "DELETE",
     });
